+++ quantify_scheduler/device_under_test/transmon_element.py
@@ -523,6 +523,7 @@ class BasicTransmonElement(DeviceElement):
                         "acq_rotation": self.measure.acq_rotation(),
                         "acq_threshold": self.measure.acq_threshold(),
                         "num_points": self.measure.num_points(),
+                        "freq": None,
                     },
                     gate_info_factory_kwargs=[
                         "acq_channel_override",