+++ CHANGELOG.md
@@ -66,6 +66,7 @@
   - Make it possible to override gate operations with device level parameters. (!1018)
   - New schedule `cnot_as_h_cz_h` in composite_factory. (!1020)
   - Introduce `Timetag` acquisition protocol and `Timestamp` operation. (!1023)
+  - Add ability to override frequency for some gate-level operations. (!1039)
 - QuantumDevice
   - Introduce BasicSpinElement spin device element. (!1015)